Move Github-related handler stuff to own package
